Alter container path tests to expect the container name to be returned.
We've obsoleted container path, which means that setting it is ignored, and it is filled out from the container name on get.

This updates the tests to acknowledge this.

There is also some question of the actual value allowing 'PUT' has in this circumstance, beyond the possibility of allowing 'rename.