user key regen tests
With the upcoming move of the certificate service
to keypair generation taking place in erchef, a bug
was uncovered with the new management console that
resulted in user key regeneration requests not resetting
the public_key attribute. This commit adds testing for
user key regeneration and regression tests for the
scenario described.